+++ test/github-issues/970/entity/Post.ts
@@ -1,0 +1,18 @@
+import {Entity} from "../../../../src/decorator/entity/Entity";
+import {ObjectIdColumn} from "../../../../src/decorator/columns/ObjectIdColumn";
+import {ObjectID} from "../../../../src/driver/mongodb/typings";
+import {Column} from "../../../../src/decorator/columns/Column";
+
+@Entity()
+export class Post {
+
+    @ObjectIdColumn()
+    id: ObjectID;
+
+    @Column()
+    title: string;
+
+    @Column()
+    text: string;
+
+}