+++ src/metadata/EntityMetadataUtils.ts
@@ -12,7 +12,10 @@ export class EntityMetadataUtils {
         const paths: string[] = [];
         Object.keys(entity).forEach(key => {
             if (!entity[key]) return;
-            if (entity[key] instanceof Object) {
+
+            // check for function is needed in the cases when createPropertyPath used on values containg a function as a value
+            // example: .update().set({ name: () => `SUBSTR('', 1, 2)` })
+            if (entity[key] instanceof Object && !(entity[key] instanceof Function)) {
                 const subPaths = this.createPropertyPath(entity[key], key);
                 paths.push(...subPaths);
             } else {
@@ -21,6 +24,22 @@ export class EntityMetadataUtils {
             }
         });
         return paths;
+    }
+
+    /**
+     * Creates a property paths for a given entity.
+     */
+    static getPropertyPathValue(entity: ObjectLiteral, propertyPath: string) {
+        const properties = propertyPath.split(".");
+        const recursive = (object: ObjectLiteral): any => {
+            const propertyName = properties.shift();
+            const value = propertyName ? object[propertyName] : object;
+            if (properties.length)
+                return recursive(value);
+
+            return value;
+        };
+        return recursive(entity);
     }
 
 }